% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{biweight}
\alias{biweight}
\title{Biweight Function}
\usage{
biweight(u)
}
\arguments{
\item{u}{vector of numerical values}
}
\description{
The biweight function is used as a \link[KDE:kernels]{kernel}.
}
\seealso{
\code{\link[KDE:kernels]{kernels}} for more information about kernels.

Other kernels: 
\code{\link{cosine}()},
\code{\link{epanechnikov}()},
\code{\link{gaussian}()},
\code{\link{logistic}()},
\code{\link{rectangular}()},
\code{\link{sigmoidFunction}()},
\code{\link{silverman}()},
\code{\link{triangular}()},
\code{\link{tricube}()},
\code{\link{triweight}()}
}
\concept{kernels}
