% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{cosine}
\alias{cosine}
\title{Cosine Function}
\usage{
cosine(u)
}
\arguments{
\item{u}{vector of numerical values}
}
\description{
Cosine Function
}
\seealso{
\code{\link[KDE:kernels]{kernels}} for more information about kernels.

Other kernels: 
\code{\link{biweight}()},
\code{\link{epanechnikov}()},
\code{\link{gaussian}()},
\code{\link{logistic}()},
\code{\link{rectangular}()},
\code{\link{sigmoidFunction}()},
\code{\link{silverman}()},
\code{\link{triangular}()},
\code{\link{tricube}()},
\code{\link{triweight}()}
}
\concept{kernels}
