% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation.R
\name{cross_validation}
\alias{cross_validation}
\title{Cross Validation}
\source{
\itemize{\href{https://spartacus-idh.com/030.html}{Nonparametric Estimation}, Comte \code{[}2017\verb{]}, ISBN: 978-2-36693-030-6}
}
\usage{
cross_validation(
  kernel,
  samples,
  bandwidths = logarithmic_bandwidth_set(1/length(samples), 1, 10),
  subdivisions = 100L
)
}
\arguments{
\item{kernel}{kernel function as an S3 object of the class \link[KDE:Kernel]{Kernel}.}

\item{samples}{A numerical vector of observations.}

\item{bandwidths}{The bandwidth set from which the bandwidth with the least estimated risk will be selected. The \code{cross_validation} function will try to set up a suitable bandwidth set using \code{\link[KDE:logarthmic_bandwidth_set]{logarithmic_bandwidth_set}} if \code{NULL} is passed.}

\item{subdivisions}{A integer vector of length 1 used for the subdivisions parameter of the builtin R-function \code{\link[stats:integrate]{integrate}}.}
}
\value{
A numerical vector of length 1 containing the bandwidth with minimal cross validation error.
}
\description{
The Cross Validation method is used to estimate a optimal bandwidth for kernel density estimation from a given set of bandwidths.
}
\details{
This method is an implementation of the cross validation method for bandwith estimation. The aim is to minimize the risk of a kernel density estimator (KDE). \cr
For each bandwidth \code{h} given in \code{bandwidths}, \code{cross_validation} approximates the term of the risk that is dependent of the KDE.
The risk function is given by the expected value of the integrated square error of the KDE with a given bandwith and the desired density that is matching the distibution of the samples (which we are trying to estimate with the KDE).
The method then selects the bandwidth with the minimal associated risk. \cr
For more information about cross validation for bandwidth estimation, see "Nonparametric Estimation" by Fabienne Comte.
}
\seealso{
\itemize{\code{\link[KDE:goldenshluger_lepski]{Goldenshluger-Lepski method}},
\code{\link[KDE:pco_method]{ PCO method}} to see other methods for estimating bandwidths.}
\itemize{\code{\link[KDE:Kernel]{Kernel}} to see the definiotion of a kernel.}
\itemize{\code{\link[KDE:kernel_density_estimator]{Kernel Density Estimator}} to see the functionality of the Kernel density estimation.}
}
