% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_density.R
\name{is_density}
\alias{is_density}
\title{Validation if given function is probability density function.}
\usage{
is_density(d)
}
\arguments{
\item{d}{the fuction to be tested.}
}
\value{
Boolean. Returns True if d is a probability density function, FALSE if not.
Is not supossed to return an error.
}
\description{
The \code{is_density()} function is used to check if a given
function satisfies the conditions of a probability density function. It has
to be a integrable, nonnegativ numerical function with integral over the
real numbers equal to 1.
}
\examples{
TODO

}
