% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_kernel.R
\name{is_kernel}
\alias{is_kernel}
\title{Validation if kernel is a kernel function.}
\usage{
is_kernel(kernel)
}
\arguments{
\item{kernel}{the fuction to be tested.}
}
\value{
Boolean. Returns True if kernel is a kernel function, FALSE if not. Is not supossed to return an error.
}
\description{
The \code{is_kernel()} function is used for validating a \link[KDE:kernels]{kernel function}.
It has to be a integrable, numerical function with integral over R equal to one.
}
\section{Issue}{

Because of the use of the Base-R function integrate, functions that are almost everywhere zero can be kernels, but will not be detected.
For example: kernels that are shifted with a very small bandwidth.
}

\examples{
is_kernel(gaussian)
no_kernel <- function(x) 1
is_kernel(no_kernel)

}
