% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kde-package.R
\docType{package}
\name{kde-package}
\alias{kde}
\alias{kde-package}
\title{kde: Kernel Density Estimation}
\description{
Various tools relating to Kernel Density Estimation.
}
\author{
\strong{Maintainer}: Tobias Neuschäfer \email{t.neuschaefer@stud.uni-heidelberg.de}

Authors:
\itemize{
  \item Thomas Frais
  \item Maren Heinemann
}

}
\keyword{internal}
