% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_estimator.R
\name{kernel_estimator}
\alias{kernel_estimator}
\title{Construct a Kernel Density Estimator}
\source{
Comte, F.: Nonparametric Esimation. Spartacus-Idh (2017)
}
\usage{
kernel_estimator(x, kernel = stats::dnorm, bandwith = 1)
}
\arguments{
\item{x}{numeric vector of the observation sample}

\item{kernel}{kernel function, default is Gaussian kernel}

\item{bandwith}{non-negative numeric scalar, the bandwidth of the estimator
default is 1}
}
\value{
A function which estimates the probability density function of the
sample \code{X}
}
\description{
\code{kernel_estimator} creates a kernel density estimator using
the provided sample, kernel and bandwidth
}
\details{
The function given by \code{kernel_estimator} is vectorized
}
\examples{
X <- rnorm(100)

f1 <- kernel_estimator(X, dnorm, 1)
f2 <- kernel_estimator(X, dnorm, 0.5)
f3 <- kernel_estimator(X, dnorm, 0.1)

a <- min(X)
b <- max(X)
ab <- seq(a, b, length.out = 100)

plot(ab, dnorm(ab), type = "l", xlab = "x", ylab = "density")
lines(ab, f1(ab), col = "red")
lines(ab, f2(ab), col = "blue")
lines(ab, f3(ab), col = "green")

legend("topleft",
  col = c("black", "red", "blue", "green"),
  legend = c("true", "h = 1", "h = 0.5", "h = 0.1"), pch = "l"
)

}
\seealso{
\code{\link{cross_validation}} and \code{\link{pco_method}} for
automatic bandwidth selection methods
}
