% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_transform.R
\name{kernel_transform}
\alias{kernel_transform}
\title{Shifting kernels to a coefficient and applying a bandwidth.}
\usage{
kernel_transform(kernel, sample, h)
}
\arguments{
\item{kernel}{a kernel object.}

\item{sample}{the observation/coefficient as a numerical scalar.}

\item{h}{the bandwidth as a numerical scalar.}
}
\value{
the transformed, i.e. shifted and scaled kernel function.
}
\description{
The \code{kernel_transform} is shifting and applying a bandwidth on a kernel function.
}
\examples{
# shifting the gaussian kernel
shifted_gaussian <- kernelTransform(gaussian, 2, 1)
# stretching the gaussian kernel
stretched_gaussian <- kernelTransform(gaussian, 0, 2)
x <- seq(from = -5, to = 5, length.out = 1000)

plot(x, gaussian$fun(x),
     xlim=c(-5,5), ylim=c(0,1),
     main="Kernels", xlab="", ylab="",
     col="black", type="l")
lines(x, shifted_gaussian(x), col="red")
lines(x, stretched_gaussian(x), col="blue")
legend("topright",
       legend=c("gaussian", "shifted_gaussian","stretched_gaussian"),
       col=c("black","red", "blue"), lty=1, cex=0.8)

}
