% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_density_estimator.R
\name{kernelDensityEstimator}
\alias{kernelDensityEstimator}
\title{Construct a kernel density estimator}
\usage{
kernelDensityEstimator(kernel, samples, bandwidth = 1)
}
\arguments{
\item{kernel}{The (vectorised) kernel function to use for the construction of the estimator satisfying \link{is_kernel}.}

\item{samples}{A numerical vector to base the construction of the estimator on.}

\item{bandwidth}{A non-negative numeric value to use as the bandwidth for the kernel.}
}
\value{
The constructed kernel density estimator. Therefore a function vectorised in its evaluation point returning the estimator at these points.

TODO
}
\description{
The kernelDensityEstimator function factory takes observations, returning
the corresponding kernel density estimator.
}
\seealso{
\itemize{
\item \code{\link[=kernels]{kernels()}} for a list of already implemented kernels.
\item \code{\link[=is_kernel]{is_kernel()}} to validate cusotm kernel functions.
}
}
