% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{kernels}
\alias{kernels}
\title{Kernel functions.}
\arguments{
\item{u}{vector containing numerical values.}
}
\value{
returning evaluation of the kernel function in u.
}
\description{
Built-in kernel functions.
}
\details{
{
A kernel is a real valued, integrable function, such that its integral over the real numbers equals one.
The built-in kernel functions are vectorised.
\describe{
\strong{List of built-in kernels:}
\item{\code{rectangular}}
\item{\code{triangular}}
\item{\code{epanechnikov}}
\item{\code{biweight}}
\item{\code{triweight}}
\item{\code{tricube}}
\item{\code{gaussian}}
\item{\code{cosine}}
\item{\code{logistic}}
\item{\code{sigmoidFunction}}
\item{\code{silverman}}
}
}
}
\examples{
x <- seq(from = -4, to = 4, length.out = 1000)
plot(x, gaussian(x),
     xlim=c(-5,5), ylim=c(0,1),
     main="Kernels", xlab="", ylab="",
     col="black", type="l")
lines(x,triangular(x), col="red")
lines(x, rectangular(x), col="blue")
legend("topright",
       legend=c("gaussian", "triangular","rectangular"),
       col=c("black","red", "blue"), lty=1, cex=0.8)

}
