% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logarithmic_bandwidth_set.R
\name{logarithmic_bandwidth_set}
\alias{logarithmic_bandwidth_set}
\title{Construct a logarithmic bandwidth set}
\usage{
logarithmic_bandwidth_set(from, to, length.out)
}
\arguments{
\item{from}{the starting value of the sequence (of length 1).}

\item{to}{the end value of the sequence (of length 1).}

\item{length.out}{desired length of the sequence. A non-negative number,
which will be rounded up if fractional (of length 1).}
}
\value{
A logarithmically spaced sequence of \code{length.out} elements
between \code{from} and \code{to}.
}
\description{
\code{logarithmic_bandwidth_set} returns a logarithmically
scaled sequence of \code{length.out} elements between \code{from} and
\code{to}.
}
\details{
Numerical inputs should all be \link{finite} (that is, not infinite,
\link{NaN} or NA). The \code{from} and \code{to} arguments must be strictly positive.
}
\seealso{
The method \link{seq} (for linearly scaled sequences).
}
