% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pco_method.R
\name{pco_crit}
\alias{pco_crit}
\title{Penalized Comparison To Overfitting criterion calculation}
\usage{
pco_crit(
  kernel,
  samples,
  H_n = logarithmic_bandwidth_set(1/length(samples), 1, 10),
  lambda = 1,
  subdivisions = 100L
)
}
\arguments{
\item{kernel}{kernel function as an S3 object of the class \link[KDE:Kernel]{Kernel}.}

\item{samples}{A numerical vector of observations to base the construction of the estimator.}

\item{H_n}{The bandwidth set from which the bandwidth with the least risk according to the PCO criterion will be derived. The pco_methdod function will try to set up a suitable bandwidth set if \code{NULL} is passed.}

\item{lambda}{A tuning parameter. It has to be a numerical value with length 1. The criterion blows up for lambda < 0, therefore the optimal value for lambda is a positiv real number. The recommendation is to set lambda = 1.}

\item{subdivisions}{A integer vector of length 1 used for the subdivisions parameter of the builtin R-function \code{\link[stats:integrate]{integrate}}.}
}
\value{
A list containing the bandwidth set and associated risk values according to the PCO criterion.
}
\description{
pco_crit calculates the associated risk for each bandwidth in a given bandwidth set.
}
\details{
For more information about the Penalized Comparison to Overfitting method, see \link{pco_method}.
}
\seealso{
\code{\link[KDE:pco_method]{pco_method}} for more information about the PCO method.

#' @include kernel.R
}
