% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pco_method.R
\name{pco_method}
\alias{pco_method}
\title{Penalized Comparison to Overfitting}
\usage{
pco_method(kernel, samples, H_n = NULL, lambda = 1, subdivisions = 100L)
}
\description{
Penalized Comparison to Overfitting
}
