% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rejection_sampling.R
\name{rejection_sample}
\alias{rejection_sample}
\title{Title}
\usage{
rejection_sample(
  sample_density,
  distribution,
  density,
  M = NULL,
  lower = NULL,
  upper = NULL
)
}
\arguments{
\item{upper}{}

\item{n}{}
}
\description{
Title
}
\examples{
Title

}
