% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builtin_kernels.R
\docType{data}
\name{sigmoid}
\alias{sigmoid}
\title{Sigmoid Function}
\format{
An object of class \link[KDE:Kernel]{Kernel} (inherits from IntegrableFunction) of length 2:
\itemize{
\item sigmoid function \code{fun}
\item support = \code{c(-20,20)}
A function is sufficiently small out of a compact space, such that the numeric integration outside of this space will be zero.
\item subdivisions = \code{100L}.
}
}
\usage{
\code{sigmoid$fun(u)}
}
\arguments{
\item{u}{vector of numerical values}
}
\description{
The sigmoid function is used as the function for an S3 object \code{sigmoid} of class \link[KDE:Kernel]{Kernel}.
}
\seealso{
\code{\link[KDE:Kernel]{kernels}} for more information about kernels.

Other kernels: 
\code{\link{biweight}},
\code{\link{cosine}},
\code{\link{epanechnikov}},
\code{\link{gaussian}},
\code{\link{logistic}},
\code{\link{rectangular}},
\code{\link{silverman}},
\code{\link{triangular}},
\code{\link{tricube}},
\code{\link{triweight}}
}
\concept{kernels}
\keyword{datasets}
