% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zbuiltin_kernels.R
\docType{data}
\name{silverman}
\alias{silverman}
\title{Silverman Function}
\format{
An object of class \code{Kernel} (inherits from \code{IntegrableFunction}) of length 2.
}
\usage{
silverman
}
\arguments{
\item{u}{vector of numerical values}
}
\description{
The silverman function is used as a \link[KDE:Kernel]{kernel}.
}
\seealso{
\code{\link[KDE:Kernel]{kernels}} for more information about kernels.

Other kernels: 
\code{\link{biweight}},
\code{\link{cosine}},
\code{\link{epanechnikov}},
\code{\link{gaussian}},
\code{\link{logistic}},
\code{\link{rectangular}},
\code{\link{sigmoidFunction}},
\code{\link{triangular}},
\code{\link{tricube}},
\code{\link{triweight}}
}
\concept{kernels}
\keyword{datasets}
