% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builtin_kernels.R
\docType{data}
\name{triangular}
\alias{triangular}
\title{Triangular Function}
\format{
An object of class \link[KDE:Kernel]{Kernel} (inherits from IntegrableFunction) of length 2:
\itemize{
\item triangular function \code{fun}
\item support = \code{c(-1,1)}
\item subdivisions = \code{100L}.
}
}
\usage{
\code{triangular$fun(u)}
}
\arguments{
\item{u}{vector of numerical values}
}
\value{
returning evaluation of the kernel function in u.
}
\description{
The triangular function is used as the function for an S3 object \code{triangular} of class \link[KDE:Kernel]{Kernel}.
}
\seealso{
\code{\link[KDE:Kernel]{Kernel}} for more information about kernels.

Other kernels: 
\code{\link{biweight}},
\code{\link{cosine}},
\code{\link{epanechnikov}},
\code{\link{gaussian}},
\code{\link{logistic}},
\code{\link{rectangular}},
\code{\link{sigmoid}},
\code{\link{silverman}},
\code{\link{tricube}},
\code{\link{triweight}}
}
\concept{kernels}
\keyword{datasets}
