% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{tricube}
\alias{tricube}
\title{Tricube Function}
\usage{
tricube(u)
}
\arguments{
\item{u}{vector of numerical values}
}
\description{
Tricube Function
}
\seealso{
\code{\link[KDE:kernels]{kernels}} for more information about kernels.

Other kernels: 
\code{\link{biweight}()},
\code{\link{cosine}()},
\code{\link{epanechnikov}()},
\code{\link{gaussian}()},
\code{\link{logistic}()},
\code{\link{rectangular}()},
\code{\link{sigmoidFunction}()},
\code{\link{silverman}()},
\code{\link{triangular}()},
\code{\link{triweight}()}
}
\concept{kernels}
