% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builtin_kernels.R
\docType{data}
\name{triweight}
\alias{triweight}
\title{Triweight Function}
\format{
An object of class \link[KDE:Kernel]{Kernel} (inherits from IntegrableFunction) of length 2:
\itemize{
\item triweight function \code{fun}
\item support = \code{c(-1,1)}
\item subdivisions = \code{100L}.
}
}
\usage{
\code{triweight$fun(u)}
}
\arguments{
\item{u}{vector of numerical values}
}
\description{
The triweight function is used as the function for an S3 object \code{triweight} of class \link[KDE:Kernel]{Kernel}.
}
\seealso{
\code{\link[KDE:Kernel]{kernels}} for more information about kernels.

Other kernels: 
\code{\link{biweight}},
\code{\link{cosine}},
\code{\link{epanechnikov}},
\code{\link{gaussian}},
\code{\link{logistic}},
\code{\link{rectangular}},
\code{\link{sigmoid}},
\code{\link{silverman}},
\code{\link{triangular}},
\code{\link{tricube}}
}
\concept{kernels}
\keyword{datasets}
