% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel.R
\name{validate_Kernel}
\alias{validate_Kernel}
\title{Validation if object is a kernel function.
TODO: umschreiben/mit Kernel zusammenfassen}
\usage{
validate_Kernel(obj)
}
\arguments{
\item{object}{the object to be tested.}
}
\value{
Boolean. Returns True if object is a kernel function, FALSE if not. Is not supossed to return an error.
}
\description{
The \code{is_kernel()} function is used for validating a \link[KDE:Kernel]{Kernel}.
It has to be a integrable, numerical function with integral over R equal to one.
}
\details{
Hence a Kernel is a subclass of the \code{\link[KDE:IntegrableFunction]{IntegrableFunctions}},
it makes use of \link{validate_IntegrableFunction}.
In addition, it will check wether the object is of class \link[KDE:Kernel]{Kernel} and if the integral over its support equals one.
}
\section{Issue}{

Because of the use of the Base-R function integrate, functions that are almost everywhere zero can be kernels, but will not be detected.
For example: kernels that are shifted with a very small bandwidth.
}

\examples{
is_kernel(gaussian)
no_kernel <- function(x) 1
is_kernel(no_kernel)

}
