% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builtin_kernels.R
\docType{data}
\name{biweight}
\alias{biweight}
\title{Biweight Function}
\format{
An object of class \link[KDE:Kernel]{Kernel} (inherits from IntegrableFunction) of length 2:
\itemize{
\item biweight function \code{fun}
\item support = \code{c(-1,1)}.
}
}
\usage{
\code{biweight$fun(u)}
}
\arguments{
\item{u}{vector of numerical values}
}
\value{
returning evaluation of the kernel function in u.
}
\description{
The biweight function is used as the function for an S3 object \code{biweight} of class \link[KDE:Kernel]{Kernel}.
}
\seealso{
\code{\link[KDE:Kernel]{Kernel}} for more information about kernels.

Other kernels: 
\code{\link{cosine}},
\code{\link{epanechnikov}},
\code{\link{gaussian}},
\code{\link{logistic}},
\code{\link{rectangular}},
\code{\link{sigmoid}},
\code{\link{silverman}},
\code{\link{triangular}},
\code{\link{tricube}},
\code{\link{triweight}}
}
\concept{kernels}
\keyword{datasets}
