% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zbuiltin_kernels.R
\docType{data}
\name{biweight}
\alias{biweight}
\title{Biweight Function}
\format{
An object of class \code{Kernel} (inherits from \code{IntegrableFunction}) of length 2.
}
\usage{
biweight
}
\arguments{
\item{u}{vector of numerical values}
}
\description{
The biweight function is used as a \link[KDE:Kernel]{kernel}.
}
\seealso{
\code{\link[KDE:Kernel]{kernels}} for more information about kernels.

Other kernels: 
\code{\link{cosine}},
\code{\link{epanechnikov}},
\code{\link{gaussian}},
\code{\link{logistic}},
\code{\link{rectangular}},
\code{\link{sigmoidFunction}},
\code{\link{silverman}},
\code{\link{triangular}},
\code{\link{tricube}},
\code{\link{triweight}}
}
\concept{kernels}
\keyword{datasets}
