% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation.R
\name{cross_validation}
\alias{cross_validation}
\title{Cross-Validation}
\source{
\href{https://spartacus-idh.com/030.html}{Nonparametric Estimation},
Comte [2017], ISBN: 978-2-36693-030-6
}
\usage{
cross_validation(
  kernel,
  samples,
  bandwidths = logarithmic_bandwidth_set(1/length(samples), 1, 10),
  subdivisions = 1000L
)
}
\arguments{
\item{kernel}{S3 object of class \code{\link{Kernel}}; the kernel to use for
the estimator}

\item{samples}{numeric vector; the observations.}

\item{bandwidths}{strictly positive numeric vector; the bandwidth set from
which the bandwidth with the least estimated risk will be selected.}

\item{subdivisions}{positive numeric scalar; subdivisions parameter
internally passed to \code{\link{integrate_primitive}}.}
}
\value{
The estimated optimal bandwidth contained in the bandwidth set.
}
\description{
The cross-validation method is used to estimate an optimal
bandwidth for kernel density estimation from a given set of bandwidths.
}
\details{
Cross-validation aims to minimize the mean integrated squared error
(MISE) of a kernel density estimator. The MISE is defined as the
expectation of the squared L2-Norm of the difference between estimator and
(unknown) true density.

For each bandwidth \code{h} given in \code{bandwidths},
\code{cross_validation} approximates the estimator-dependent part of the
risk. The method then selects the bandwidth with the minimal associated
risk.
}
\seealso{
\code{\link{kernel_density_estimator}} for more information about
kernel density estimators, \code{\link{pco_method}} and
\code{\link{goldenshluger_lepski}} for more automatic bandwidth-selection
algorithms.
}
