% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation.R
\name{cross_validation}
\alias{cross_validation}
\title{Cross-Validation}
\source{
Comte, F.: Nonparametric Esimation. Spartacus-Idh (2017)
}
\usage{
cross_validation(
  x,
  kernel = stats::dnorm,
  n = 40L,
  N = 100L,
  built_in = c("gaussian", "epanechnikov", "rectangular", "triangular", "biweight",
    "silverman"),
  na.rm = FALSE
)
}
\arguments{
\item{x}{numeric vector of the observation sample}

\item{kernel}{kernel function used for the kernel density estimation, default
is Gaussian kernel}

\item{n}{number of bandwidths to be optimized from. \code{cross_validation}
selects a bandwidth contained in  (1/n, 2/n, ..., 1)}

\item{N}{number of subdivisions used in discretization of integrals}

\item{built_in}{choose one of the built-in kernels instead of providing one yourself}

\item{na.rm}{logical; if TRUE, missing values will be removed from x}
}
\value{
The estimate optimal bandwidth
}
\description{
\code{cross_validation} estimates an optimal bandwidth
for kernel density estimation using the cross-validation method
}
\details{
The cross-validation method tries to minimize the mean squared
integrated error (MISE) of a kernel density estimator

\code{cross_validation} approximates the estimator-dependent part of the
risk and selects the bandwidth with the minimal risk
}
\examples{

x <- stats::rnorm(100)
h <- cross_validation(x, kernel = stats::dnorm)
f <- kernel_estimator(x, kernel = stats::dnorm, bandwidth = h)

a <- min(x)
b <- max(x)
ab <- seq(a, b, length.out = 100)

plot(ab, stats::dnorm(ab), type = "l")
lines(ab, f(ab), col = "red")
legend("topleft",
  legend = c("true", "estimated h"), col = c("black", "red"),
  pch = "|"
)

}
\seealso{
\code{\link{kernel_estimator}} for more information about kernel
density estimation

\code{\link{goldenshluger_lepski}} and \code{\link{pco_method}} for other automatic bandwidth-selection
algorithms
}
