% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_estimator.R
\name{epanechnikov}
\alias{epanechnikov}
\title{Epanechnikov kernel}
\usage{
epanechnikov(u)
}
\arguments{
\item{u}{real number}
}
\value{
value of Epanechnikov kernel for u
}
\description{
Epanechnikov kernel
}
