% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrable_function.R
\name{IntegrableFunction}
\alias{IntegrableFunction}
\title{Integrable-Function Classes}
\usage{
IntegrableFunction(fun, support)
}
\description{
The S3 class \code{'IntegrableFunction'} serves as superclass for the S3 classes \code{'Density'} and \code{'Kernel'}.
}
