% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrable_function.R
\name{IntegrableFunction}
\alias{IntegrableFunction}
\title{Integrable-Function Classes}
\usage{
IntegrableFunction(fun, support = NULL)
}
\arguments{
\item{fun}{an \code{R} function taking a single numeric argument and
returning a numeric vector of the same length: see 'Details'.}

\item{support}{a numerical vector of length 2 containing the lower- and
upperbound in the first and second entry respectively.
\code{IntegrableFunction} will try to find bounds on the support itself if
\code{NULL} is passed: see 'Details'.}
}
\description{
The S3 class \code{IntegrableFunction} serves as superclass for
the S3 classes \code{Density} and \code{Kernel}. The constructor tries to
construct a valid \code{IntegrableFunction}.
}
\details{
In this package the S3 class \code{IntegrableFunction} exists to ensure
the integrability of \code{R} functions. Valid \code{IntegrableFunction} objects are build
on lists and contain two names entries:
\itemize{
\item \strong{\code{fun}} is an \code{R} function taking a single numeric argument and
returning a numeric vector of the same length. This function should return
zero outside of the interval given in the \code{support} entry.
\item \strong{\code{support}} is a numeric vector of length 2 containing a lower- and
upperbound for the support of the function stored in \code{fun} in its first
and second entry respectively. In particular the values \code{-Inf} and
\code{Inf} are allowed.
}

The constructor \code{IntegrableFunction} tries to construct an
\code{IntegrableFunction} object based on the provided arguments. In
particular it assigns the correct class attribute and verifies the correct
structure of the \code{fun} and \code{support} entries, such that the returned
object passes a test using \code{\link[=validate_IntegrableFunction]{validate_IntegrableFunction()}}.
}
