% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_kernel.R
\name{is_kernel}
\alias{is_kernel}
\title{Validation if kernel is a kernel function.}
\usage{
is_kernel(kernel)
}
\arguments{
\item{kernel}{the fuction to be tested.}
}
\value{
Boolean. Returns True if kernel is a kernel function, FALSE if not. Is not supossed to return an error.
}
\description{
Validation if kernel is a kernel function.
}
