% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kde-package.R
\docType{package}
\name{kde-package}
\alias{kde}
\alias{kde-package}
\title{kde: Kernel Density Estimation}
\description{
This package contains a method for Kernel Density Estimation (KDE) as well as three different methods to automatically select an approximatively optimal bandwidth for KDE. You can also generate samples for a given probability density using Rejection Sampling.
}
\author{
\strong{Maintainer}: Tobias Neuschäfer \email{t.neuschaefer@stud.uni-heidelberg.de}

Authors:
\itemize{
  \item Thomas Frais
  \item Maren Heinemann
}

}
\keyword{internal}
