% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel.R
\name{Kernel}
\alias{Kernel}
\title{Kernel}
\usage{
Kernel(fun, support)
}
\arguments{
\item{u}{vector containing numerical values.}
}
\value{
returning evaluation of the kernel function in u.
}
\description{
Built-in kernels are S3 Objects, holding a function and its support.
}
\details{
{
A kernel is a real valued, integrable function, such that its integral over the real numbers equals one.
The built-in kernel functions are vectorised.
\describe{
\strong{List of built-in kernels:}
\item{\code{\link[KDE:rectangular]{rectangular}}}
\item{\code{\link[KDE:triangular]{triangular}}}
\item{\code{\link[KDE:epanechnikov]{epanechnikov}}}
\item{\code{\link[KDE:biweight]{biweight}}}
\item{\code{\link[KDE:triweight]{triweight}}}
\item{\code{\link[KDE:tricube]{tricube}}}
\item{\code{\link[KDE:gaussian]{gaussian}}}
\item{\code{\link[KDE:cosine]{cosine}}}
\item{\code{\link[KDE:logistic]{logistic}}}
\item{\code{\link[KDE:sigmoidFunction]{sigmoidFunction}}}
\item{\code{\link[KDE:silverman]{silverman}}}
}
}
}
\examples{
x <- seq(from = -4, to = 4, length.out = 1000)
plot(x, gaussian$fun(x),
     xlim=c(-5,5), ylim=c(0,1),
     main="Kernels", xlab="", ylab="",
     col="black", type="l")
lines(x,triangular$fun(x), col="red")
lines(x, rectangular$fun(x), col="blue")
legend("topright",
       legend=c("gaussian", "triangular","rectangular"),
       col=c("black","red", "blue"), lty=1, cex=0.8)

}
\seealso{
\code{\link[KDE:validate_Kernel]{validate_Kernel}
}
