% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_density_estimator.R
\name{kernel_density_estimator}
\alias{kernel_density_estimator}
\title{Construct a kernel density estimator}
\source{
Nonparametric Estimation, Comte [2017], ISBN: 978-2-36693-030-6
}
\usage{
kernel_density_estimator(kernel, samples, bandwidth = 1, subdivisions = 100L)
}
\arguments{
\item{kernel}{a kernel as S3 object of the class \link{Kernel}.}

\item{samples}{a numerical vector of observations.}

\item{bandwidth}{a non-negative numeric value to use as the bandwidth for the
estimator.}

\item{subdivisions}{a integer vector of length 1 used for the subdivisions
parameter of the builtin R-function \code{\link{integrate}}.}
}
\value{
The estimator is returned as S3 object of class
\code{\link{IntegrableFunction}}.
}
\description{
The \code{kernel_density_estimator} function builds a kernel
density estimator using the provided kernel, bandwidth and samples.
}
\details{
The validation of the returned estimator as
\code{\link{IntegrableFunction}} relies on the builtin function
\code{\link{integrate}}, which requires a \code{subdivisions} argument.
Integration using a larger number of subdivisions will increase the
runtime. In contrast too few subdivisions may result in a runtime error.

For more information about kernel density estimaton, see "Nonparametric
Estimation" by Fabienne Comte.
}
\seealso{
\code{\link{Kernel}} for more information about kernels,
\code{\link[KDE:pco_method]{PCO}},
\code{\link[KDE:cross_validation]{Cross-Validation}} and
\code{\link[KDE:goldenshluger_lepski]{Goldenshluger-Lepski}} for
automatic bandwidth-selection algorithms.
}
