% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_density_estimator.R
\name{kernel_density_estimator}
\alias{kernel_density_estimator}
\title{Construct a kernel density estimator}
\usage{
kernel_density_estimator(kernel, samples, bandwidth = 1, subdivisions = 100L)
}
\arguments{
\item{kernel}{The (vectorised) kernel function to use for the construction of
the estimator satisfying \link{is_kernel}.}

\item{samples}{A numerical vector to base the construction of the estimator
on.}

\item{bandwidth}{A non-negative numeric value to use as the bandwidth for the
kernel.}
}
\value{
The constructed kernel density estimator. Therefore a function
vectorised in its evaluation point returning the estimator at these points.
}
\description{
The kernelDensityEstimator function factory takes observations,
returning the corresponding kernel density estimator.
}
\seealso{
\itemize{
\item \code{\link[=kernels]{kernels()}} for a list of already implemented kernels. *
\code{\link[=validate_kernel]{validate_kernel()}} to validate cusotm kernel functions.
}
}
