% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_density_estimator.R
\name{kernel_density_estimator}
\alias{kernel_density_estimator}
\title{Construct a Kernel Density Estimator}
\source{
Nonparametric Estimation, Comte [2017], ISBN: 978-2-36693-030-6
}
\usage{
kernel_density_estimator(kernel, samples, bandwidth = 1, subdivisions = 1000L)
}
\arguments{
\item{kernel}{a kernel as S3 object of the class \link{Kernel}.}

\item{samples}{numeric vector; the observations.}

\item{bandwidth}{non-negative numeric scalar; the bandwidth for the
estimator.}

\item{subdivisions}{positive numeric scalar; subdivisions parameter
internally passed to \code{\link{integrate_primitive}}.}
}
\value{
The estimator as S3 object of class \code{\link{IntegrableFunction}}.
}
\description{
The \code{kernel_density_estimator} function builds a kernel
density estimator using the provided kernel, bandwidth and samples.
}
\details{
The validation of the returned estimator as
\code{\link{IntegrableFunction}} relies on the function
\code{integrate_primitive}, thus the \code{subdivisions} parameter.
}
\seealso{
\code{\link{Kernel}} for more information about kernels,
\code{\link{pco_method}}, \code{\link{cross_validation}} and
\code{\link{goldenshluger_lepski}} for automatic bandwidth-selection
algorithms.
}
