% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_transform.R
\name{kernel_transform}
\alias{kernel_transform}
\title{Shifting kernels to a coefficient and applying a bandwidth.}
\usage{
kernel_transform(kernel, sample, h, subdivisions = 100L)
}
\arguments{
\item{kernel}{a kernel object.}

\item{sample}{the observation/coefficient as a numerical scalar.}

\item{h}{the bandwidth as a numerical scalar.}

\item{subdivisions}{a integer vector of length 1, used for the subdivisions parameter for the function \code{\link[stats::integrate]{integrate}}.}
}
\value{
The transformed (shifted and scaled) kernel as valid S3 object of
class \code{Kernel}.
}
\description{
The \code{kernel_transform} is shifting and applying a bandwidth to a kernel.
}
\examples{
# shifting the gaussian kernel
shifted_gaussian <- kernel_transform(gaussian, 2, 1)
# stretching the gaussian kernel
stretched_gaussian <- kernel_transform(gaussian, 0, 2)
x <- seq(from = -5, to = 5, length.out = 1000)

plot(x, gaussian$fun(x),
     xlim=c(-5,5), ylim=c(0,1),
     main="Kernels", xlab="", ylab="",
     col="black", type="l")
lines(x, shifted_gaussian$fun(x), col="red")
lines(x, stretched_gaussian$fun(x), col="blue")
legend("topright",
       legend=c("gaussian", "shifted_gaussian","stretched_gaussian"),
       col=c("black","red", "blue"), lty=1, cex=0.8)

}
\seealso{
\code{\link[KDE:Kernel]{Kernel}}
\code{\link[KDE:validate_Kernel]{validate_Kernel}}
\code{\link[KDE:IntegrableFunction]{IntegrableFunction}}
}
