% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_transform.R
\name{kernel_transform}
\alias{kernel_transform}
\title{Shifting Kernels to a Coefficient and applying a Bandwidth.}
\usage{
kernel_transform(kernel, sample, bandwidth, subdivisions = 1000L)
}
\arguments{
\item{kernel}{a kernel as S3 object of the class \link{Kernel}.}

\item{sample}{numeric scalar; the observation.}

\item{bandwidth}{numeric scalar; the bandwidth.}

\item{subdivisions}{positive numeric scalar; subdivisions parameter
internally passed to \code{\link{integrate_primitive}}.}
}
\value{
The transformed (shifted and scaled) kernel as valid S3 object of
class \code{Kernel}.
}
\description{
\code{kernel_transform} is shifting and applying a bandwidth to a kernel.
}
\details{
The validation of the returned estimator as
\code{\link{IntegrableFunction}} relies on the function
\code{\link{integrate_primitive}}, thus the \code{subdivisions} parameter.
}
\examples{
# shifting the gaussian kernel
shifted_gaussian <- kernel_transform(gaussian, 2, 1)
# stretching the gaussian kernel
stretched_gaussian <- kernel_transform(gaussian, 0, 2)
x <- seq(from = -5, to = 5, length.out = 1000)

plot(x, gaussian$fun(x),
     xlim=c(-5,5), ylim=c(0,1),
     main="Kernels", xlab="", ylab="",
     col="black", type="l")
lines(x, shifted_gaussian$fun(x), col="red")
lines(x, stretched_gaussian$fun(x), col="blue")
legend("topright",
       legend=c("gaussian", "shifted_gaussian","stretched_gaussian"),
       col=c("black","red", "blue"), lty=1, cex=0.8)

}
\seealso{
\code{\link{Kernel}} for more information about kernels.
}
