% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logarithmic_bandwidth_set.R
\name{logarithmic_bandwidth_set}
\alias{logarithmic_bandwidth_set}
\title{Construct a Logarithmic Bandwidth Set}
\usage{
logarithmic_bandwidth_set(from, to, length.out)
}
\arguments{
\item{from}{strictly positive numeric scalar; the starting value of the
sequence.}

\item{to}{strictly positive numeric scalar; the end value of the sequence.}

\item{length.out}{non-negative numeric scalar; desired length of the sequence
(will be rounded up if fractional).}
}
\value{
A logarithmically spaced sequence of \code{length.out} elements
between \code{from} and \code{to}.
}
\description{
\code{logarithmic_bandwidth_set} returns a logarithmically
scaled sequence of \code{length.out} elements between \code{from} and
\code{to}.
}
\details{
Numerical inputs should all be \link[base:is.finite]{finite} (that is,
not infinite, \code{NaN} or \code{NA}).
}
\seealso{
\code{\link{seq}} for linearly scaled sequences.
}
