% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pco_method.R
\name{pco_method}
\alias{pco_method}
\title{Penalized Comparison to Overfitting}
\source{
\href{https://arxiv.org/abs/1607.05091v2}{Estimator selection: a new
method with applications to KDEs}, Lacour [2017]

\href{https://arxiv.org/abs/1902.01075}{Numerical performance of
PCO for multivariate KDEs}, Varet [2019]
}
\usage{
pco_method(
  kernel,
  samples,
  bandwidths = logarithmic_bandwidth_set(1/length(samples), 1, 10),
  lambda = 1,
  subdivisions = 100L
)
}
\arguments{
\item{kernel}{S3 object of class \code{\link{Kernel}}; the kernel to use for
the estimator}

\item{samples}{numeric vector; the observations.}

\item{bandwidths}{strictly positive numeric vector; the bandwidth set from
which the bandwidth with the least estimated risk will be selected.}

\item{lambda}{positive numeric scalar; a tuning parameter.}

\item{subdivisions}{positive numeric scalar; subdivisions parameter
internally passed to \code{\link{integrate_primitive}}.}
}
\value{
The estimated optimal bandwidth contained in the bandwidth set.
}
\description{
The PCO method is used to estimate an optimal bandwidth for
kernel density estimation from a given set of bandwidths.
}
\details{
The PCO method aims to minimize an upper bound for the mean
integrated squared error (MISE) of a kernel density estimator. The MISE is
defined as the expectation of the squared L2-Norm of the difference between
estimator and (unknown) true density.

\code{pco_method} internally uses a criterion function to calculate the PCO
criterion value, approximating the risk. Subsequently the bandwidth with
the minimal criterion value is selected.

The popular bias-/variance-decomposition is used. The bias term still
depends on the unknown density. Thus, a comparison of the estimator with an
associated bandwidth to the overfitting one, namely the estimator with the
smallest bandwidth, is used to estimate the bias term itself.

Further a penalty term is computed as the sum of two variances,
particularly the variance of the risk decomposition and the variance of the
bias term estimation. During the calculation the tuning parameter
\code{lambda} is used. The recommended value for \code{lambda} is 1.

The PCO criterion is given by the sum of the comparison to overfitting and
the penalty term, thus the procedure tries to find a balance between those
terms. Therefore, it is comprehensible why this method is called penalized
comparison to overfitting.

For more information see the linked papers below.
}
\seealso{
\code{\link{kernel_density_estimator}} for more information about
kernel density estimators, \code{\link{cross_validation}} and
\code{\link{goldenshluger_lepski}} for more automatic bandwidth-selection
algorithms.
}
