% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builtin_kernels.R
\docType{data}
\name{rectangular}
\alias{rectangular}
\title{Rectangular Function}
\format{
An object of class \link[KDE:Kernel]{Kernel} (inherits from IntegrableFunction) of length 2:
\itemize{
\item rectangular function \code{fun}
\item support = \code{c(-1,1)}.
}
}
\usage{
\code{rectangular$fun(u)}
}
\arguments{
\item{u}{vector of numerical values}
}
\value{
returning evaluation of the kernel function in u.
}
\description{
The rectangular function is used as the function for an S3 object \code{rectangular} of class \link[KDE:Kernel]{Kernel}.
}
\seealso{
\code{\link[KDE:Kernel]{Kernel}} for more information about kernels.

Other kernels: 
\code{\link{biweight}},
\code{\link{cosine}},
\code{\link{epanechnikov}},
\code{\link{gaussian}},
\code{\link{logistic}},
\code{\link{sigmoidFunction}},
\code{\link{silverman}},
\code{\link{triangular}},
\code{\link{tricube}},
\code{\link{triweight}}
}
\concept{kernels}
\keyword{datasets}
