% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_estimator.R
\name{rectangular}
\alias{rectangular}
\title{Rectangular Kernel}
\usage{
rectangular(u)
}
\arguments{
\item{u}{real number}
}
\value{
value of Rectangular kernel for u
}
\description{
Rectangular Kernel
}
