% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builtin_kernels.R
\docType{data}
\name{silverman}
\alias{silverman}
\title{Silverman Function}
\format{
An object of class \link[KDE:Kernel]{Kernel} (inherits from IntegrableFunction) of length 2:
\itemize{
\item silverman function \code{fun}
\item support = \code{c(-Inf,Inf)}.
}
}
\usage{
\code{silverman$fun(u)}
}
\arguments{
\item{u}{vector of numerical values}
}
\description{
The silverman function is used as the function for an S3 object \code{silverman} of class \link[KDE:Kernel]{Kernel}.
}
\seealso{
\code{\link[KDE:Kernel]{kernels}} for more information about kernels.

Other kernels: 
\code{\link{biweight}},
\code{\link{cosine}},
\code{\link{epanechnikov}},
\code{\link{gaussian}},
\code{\link{logistic}},
\code{\link{rectangular}},
\code{\link{sigmoid}},
\code{\link{triangular}},
\code{\link{tricube}},
\code{\link{triweight}}
}
\concept{kernels}
\keyword{datasets}
