% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builtin_kernels.R
\docType{data}
\name{triangular}
\alias{triangular}
\title{Triangular Kernel}
\format{
An object of S3 class \code{\link{Kernel}} with named entries
\itemize{
\item \bold{\code{fun}}: the triangular function
\item \bold{\code{support}}: \code{c(-1,1)}
\item \bold{\code{subdivisions}}: \code{1000L}.
}
}
\usage{
triangular
}
\description{
The triangular kernel is S3 object of class \code{Kernel}
provided by the \code{KDE} package.
}
\seealso{
\code{\link[KDE:Kernel]{Kernel}} for more information about kernels.

Other kernels: 
\code{\link{biweight}},
\code{\link{cosine}},
\code{\link{epanechnikov}},
\code{\link{gaussian}},
\code{\link{logistic}},
\code{\link{rectangular}},
\code{\link{sigmoid}},
\code{\link{silverman}},
\code{\link{tricube}},
\code{\link{triweight}}
}
\concept{kernels}
\keyword{datasets}
